Add ability to save remote files in Nuclide
Nuclide (https://nuclide.io/) provides the ability to edit remote files whose file path look like this: `'nuclide://hostname/folder/file.js'`
In that particular case, it doesn't make sense to try to open a local file so we can skip this file check.

Note: in case one changes of remote git branch on the server, it's possible that an Atom paneItem points to a non-existent remote file.
In that case, the plugin would re-create this file. So far, it's not a big issue since one can use the git clean feature to remove them automatically.